:recycle: api: Move mailer to src
