+++ api/src/shared/infrastructure/mailers/MailingProviderInvalidEmailError.js
