+++ api/tests/shared/unit/infrastructure/mailers/MailingProvider_test.js
@@ -1,5 +1,5 @@
-import { expect } from '../../../test-helper.js';
-import { MailingProvider } from '../../../../lib/infrastructure/mailers/MailingProvider.js';
+import { expect } from '../../../../test-helper.js';
+import { MailingProvider } from '../../../../../src/shared/infrastructure/mailers/MailingProvider.js';
 
 describe('Unit | Infrastructure | Mailers | MailingProvider', function () {
   describe('#sendEmail', function () {
