+++ api/tests/shared/unit/infrastructure/mailers/mailer_test.js
@@ -1,9 +1,10 @@
-import { expect, sinon } from '../../../test-helper.js';
-import { config } from '../../../../lib/config.js';
-import { logger } from '../../../../lib/infrastructure/logger.js';
-import { Mailer } from '../../../../lib/infrastructure/mailers/mailer.js';
-import { EmailingAttempt } from '../../../../lib/domain/models/EmailingAttempt.js';
-import { MailingProviderInvalidEmailError } from '../../../../lib/infrastructure/mailers/MailingProviderInvalidEmailError.js';
+import { sinon, expect } from '../../../../test-helper.js';
+import { config } from '../../../../../src/shared/config.js';
+import { logger } from '../../../../../src/shared/infrastructure/utils/logger.js';
+import { Mailer } from '../../../../../src/shared/infrastructure/mailers/mailer.js';
+import { EmailingAttempt } from '../../../../../src/shared/domain/models/EmailingAttempt.js';
+import { MailingProviderInvalidEmailError } from '../../../../../src/shared/infrastructure/mailers/MailingProviderInvalidEmailError.js';
+
 const { mailing } = config;
 
 describe('Unit | Infrastructure | Mailers | mailer', function () {
